+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 ChangeLog
 --------
 
+##### 1.2.0
+* Add properties : ```orientation```(default:horizontal) ， ```gravity```(default:center)
+* Remove auto register DataSetObserver. [#54](https://github.com/ongakuer/CircleIndicator/issues/54)
+
+ <b>Please change like this if you used dynamic adapter:</b>
+```java
+viewpager.setAdapter(mAdapter);
+indicator.setViewPager(viewpager);
+mAdapter.registerDataSetObserver(indicator.getDataSetObserver());
+```
+
 ##### 1.1.9  (2016-04-16)
 * Fix bug ([#51](https://github.com/ongakuer/CircleIndicator/issues/51))
 * Add Reset Adapter Sample